+++ packages/api/eslint-config-nestjs/configs/default.js
@@ -1,13 +1,12 @@
 import tseslint from 'typescript-eslint'
 import globals from 'globals'
-import eslintImportTypescript from 'eslint-plugin-import-typescript'
 
 /**
  * @type {import('eslint').Linter.Config[]}
  */
 const config = [
   {
-    files: ['**/*.ts', "lib/**/*.js"],
+    files: ['**/*.ts', 'lib/**/*.js'],
     languageOptions: {
       parser: tseslint.parser,
       parserOptions: {
@@ -25,20 +24,26 @@ const config = [
       '@typescript-eslint/naming-convention': [
         'error',
         { selector: 'default', format: ['camelCase'], leadingUnderscore: 'allow' },
-        { selector: 'variable', format: ['camelCase', 'PascalCase', 'UPPER_CASE'], leadingUnderscore: 'allow' },
+        {
+          selector: 'variable',
+          format: ['camelCase', 'PascalCase', 'UPPER_CASE'],
+          leadingUnderscore: 'allow'
+        },
         { selector: 'import', format: ['camelCase', 'PascalCase'] },
         { selector: 'objectLiteralProperty', format: null },
         { selector: 'typeLike', format: ['PascalCase'] },
         { selector: 'interface', format: ['PascalCase'] },
         { selector: 'class', format: ['PascalCase'] },
         { selector: 'enumMember', format: ['UPPER_CASE'] },
-        { selector: 'classProperty', format: ['camelCase', 'UPPER_CASE'], leadingUnderscore: 'allow' },
+        {
+          selector: 'classProperty',
+          format: ['camelCase', 'UPPER_CASE'],
+          leadingUnderscore: 'allow'
+        },
         { selector: 'function', format: ['camelCase', 'PascalCase'] },
-        { selector: 'classicAccessor', format: ['camelCase', 'PascalCase'] },
+        { selector: 'classicAccessor', format: ['camelCase', 'PascalCase'] }
       ],
-      'nonblock-statement-body-position': [
-        'error', 'beside'
-      ]
+      'nonblock-statement-body-position': ['error', 'beside']
     }
   },
   {
@@ -46,21 +51,7 @@ const config = [
     ...tseslint.configs.disableTypeChecked
   },
   {
-    ignores: [
-      'dist',
-      'node_modules',
-      'src/modules/localization/generated/i18n.generated.ts'
-    ]
-  },
-  {
-    plugins: {
-      'import-typescript': eslintImportTypescript
-    },
-    rules: {
-      'import-typescript/no-relative-parent-imports': [
-        'error', { onlyPathsImport: true }
-      ]
-    }
+    ignores: ['dist', 'node_modules', 'src/modules/localization/generated/i18n.generated.ts']
   }
 ]