+++ packages/web/design-system/src/ui/select/SelectOption.vue
@@ -29,6 +29,7 @@ const isUsingKeyboard = useIsUsingKeyboard()
 <template>
   <RekaListboxItem
     :value="props.value"
+
     :class="{
       'data-highlighted:bg-secondary-hover': isUsingKeyboard,
       'hover:bg-secondary-hover': !isUsingKeyboard,